// src/web/lib/api.ts
export class CreatorPartialDto implements ICreatorPartialDto {
    avatar_url!: string;
    created_at!: Date;
    history!: VoteDto[];
    id!: number;
    name!: string;
    platform!: Platform;
    platform_id!: string;
    platform_url!: string;
    slug!: string;
    stream_status!: StreamStatusDto;
    ticker!: string;
    updated_at!: Date;
    value!: number;

    constructor(data?: ICreatorPartialDto) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (this as any)[property] = (data as any)[property];
            }
        }
        if (!data) {
            this.history = [];
            this.stream_status = new StreamStatusDto();
        }
    }

    init(_data?: any) {
        if (_data) {
            this.avatar_url = _data["avatar_url"] !== undefined ? _data["avatar_url"] : null as any;
            this.created_at = _data["created_at"] ? new Date(_data["created_at"].toString()) : null as any;
            if (Array.isArray(_data["history"])) {
                this.history = [] as any;
                for (let item of _data["history"])
                    this.history!.push(VoteDto.fromJS(item));
            }
            else {
                this.history = null as any;
            }
            this.id = _data["id"] !== undefined ? _data["id"] : null as any;
            this.name = _data["name"] !== undefined ? _data["name"] : null as any;
            this.platform = _data["platform"] !== undefined ? _data["platform"] : null as any;
            this.platform_id = _data["platform_id"] !== undefined ? _data["platform_id"] : null as any;
            this.platform_url = _data["platform_url"] !== undefined ? _data["platform_url"] : null as any;
            this.slug = _data["slug"] !== undefined ? _data["slug"] : null as any;
            this.stream_status = _data["stream_status"] ? StreamStatusDto.fromJS(_data["stream_status"]) : new StreamStatusDto();
            this.ticker = _data["ticker"] !== undefined ? _data["ticker"] : null as any;
            this.updated_at = _data["updated_at"] ? new Date(_data["updated_at"].toString()) : null as any;
            this.value = _data["value"] !== undefined ? _data["value"] : null as any;
        }
    }

    static fromJS(data: any): CreatorPartialDto {
        data = typeof data === 'object' ? data : {};
        let result = new CreatorPartialDto();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["avatar_url"] = this.avatar_url !== undefined ? this.avatar_url : null as any;
        data["created_at"] = this.created_at ? this.created_at.toISOString() : null as any;
        if (Array.isArray(this.history)) {
            data["history"] = [];
            for (let item of this.history)
                data["history"].push(item ? item.toJSON() : null as any);
        }
        data["id"] = this.id !== undefined ? this.id : null as any;
        data["name"] = this.name !== undefined ? this.name : null as any;
        data["platform"] = this.platform !== undefined ? this.platform : null as any;
        data["platform_id"] = this.platform_id !== undefined ? this.platform_id : null as any;
        data["platform_url"] = this.platform_url !== undefined ? this.platform_url : null as any;
        data["slug"] = this.slug !== undefined ? this.slug : null as any;
        data["stream_status"] = this.stream_status ? this.stream_status.toJSON() : null as any;
        data["ticker"] = this.ticker !== undefined ? this.ticker : null as any;
        data["updated_at"] = this.updated_at ? this.updated_at.toISOString() : null as any;
        data["value"] = this.value !== undefined ? this.value : null as any;
        return data;
    }
}

/** Mirrors `TTX.App.Dto.Creators.CreatorPartialDto`. */
export interface ICreatorPartialDto {
    avatar_url: string;
    created_at: Date;
    history: VoteDto[];
    id: number;
    name: string;
    platform: Platform;
    platform_id: string;
    platform_url: string;
    slug: string;
    stream_status: StreamStatusDto;
    ticker: string;
    updated_at: Date;
    value: number;
}

/** Mirrors `TTX.App.Dto.Creators.CreatorDto`. */
export class CreatorDto extends CreatorPartialDto implements ICreatorDto {
    shares!: CreatorShareDto[];
    transactions!: CreatorTransactionDto[];

    constructor(data?: ICreatorDto) {
        super(data);
        if (!data) {
            this.shares = [];
            this.transactions = [];
        }
    }

    override init(_data?: any) {
        super.init(_data);
        if (_data) {
            if (Array.isArray(_data["shares"])) {
                this.shares = [] as any;
                for (let item of _data["shares"])
                    this.shares!.push(CreatorShareDto.fromJS(item));
            }
            else {
                this.shares = null as any;
            }
            if (Array.isArray(_data["transactions"])) {
                this.transactions = [] as any;
                for (let item of _data["transactions"])
                    this.transactions!.push(CreatorTransactionDto.fromJS(item));
            }
            else {
                this.transactions = null as any;
            }
        }
    }

    static override fromJS(data: any): CreatorDto {
        data = typeof data === 'object' ? data : {};
        let result = new CreatorDto();
        result.init(data);
        return result;
    }

    override toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        if (Array.isArray(this.shares)) {
            data["shares"] = [];
            for (let item of this.shares)
                data["shares"].push(item ? item.toJSON() : null as any);
        }
        if (Array.isArray(this.transactions)) {
            data["transactions"] = [];
            for (let item of this.transactions)
                data["transactions"].push(item ? item.toJSON() : null as any);
        }
        super.toJSON(data);
        return data;
    }
}

/** Mirrors `TTX.App.Dto.Creators.CreatorDto`. */
export interface ICreatorDto extends ICreatorPartialDto {
    shares: CreatorShareDto[];
    transactions: CreatorTransactionDto[];
}
// ...
export enum CreatorOrderBy {
    Name = "Name",
    Value = "Value",
    IsLive = "IsLive",
}
// ...
export class CreatorShareDto implements ICreatorShareDto {
    player!: PlayerPartialDto;
    quantity!: number;

    constructor(data?: ICreatorShareDto) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (this as any)[property] = (data as any)[property];
            }
        }
        if (!data) {
            this.player = new PlayerPartialDto();
        }
    }

    init(_data?: any) {
        if (_data) {
            this.player = _data["player"] ? PlayerPartialDto.fromJS(_data["player"]) : new PlayerPartialDto();
            this.quantity = _data["quantity"] !== undefined ? _data["quantity"] : null as any;
        }
    }

    static fromJS(data: any): CreatorShareDto {
        data = typeof data === 'object' ? data : {};
        let result = new CreatorShareDto();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["player"] = this.player ? this.player.toJSON() : null as any;
        data["quantity"] = this.quantity !== undefined ? this.quantity : null as any;
        return data;
    }
}

/** Mirrors `TTX.App.Dto.Creators.CreatorShareDto`. */
export interface ICreatorShareDto {
    player: PlayerPartialDto;
    quantity: number;
}
// ...
export class CreatorTransactionDto extends TransactionDto implements ICreatorTransactionDto {
    player!: PlayerPartialDto;

    constructor(data?: ICreatorTransactionDto) {
        super(data);
        if (!data) {
            this.player = new PlayerPartialDto();
        }
    }

    override init(_data?: any) {
        super.init(_data);
        if (_data) {
            this.player = _data["player"] ? PlayerPartialDto.fromJS(_data["player"]) : new PlayerPartialDto();
        }
    }

    static override fromJS(data: any): CreatorTransactionDto {
        data = typeof data === 'object' ? data : {};
        let result = new CreatorTransactionDto();
        result.init(data);
        return result;
    }

    override toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["player"] = this.player ? this.player.toJSON() : null as any;
        super.toJSON(data);
        return data;
    }
}

/** Mirrors `TTX.App.Dto.Transactions.CreatorTransactionDto`. */
export interface ICreatorTransactionDto extends ITransactionDto {
    player: PlayerPartialDto;
}
// ...
export enum OrderDirection {
    Ascending = "Ascending",
    Descending = "Descending",
}
// ...
export enum Platform {
    Twitch = "Twitch",
}
// ...
export enum PlayerOrderBy {
    Name = "Name",
    Credits = "Credits",
    Portfolio = "Portfolio",
}
// ...
export class StreamStatusDto implements IStreamStatusDto {
    ended_at?: Date | null;
    is_live!: boolean;
    started_at?: Date | null;

    constructor(data?: IStreamStatusDto) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (this as any)[property] = (data as any)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.ended_at = _data["ended_at"] ? new Date(_data["ended_at"].toString()) : null as any;
            this.is_live = _data["is_live"] !== undefined ? _data["is_live"] : null as any;
            this.started_at = _data["started_at"] ? new Date(_data["started_at"].toString()) : null as any;
        }
    }

    static fromJS(data: any): StreamStatusDto {
        data = typeof data === 'object' ? data : {};
        let result = new StreamStatusDto();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["ended_at"] = this.ended_at ? this.ended_at.toISOString() : null as any;
        data["is_live"] = this.is_live !== undefined ? this.is_live : null as any;
        data["started_at"] = this.started_at ? this.started_at.toISOString() : null as any;
        return data;
    }
}
// ...
export class VoteDto implements IVoteDto {
    creator_id!: number;
    time!: Date;
    value!: number;

    constructor(data?: IVoteDto) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (this as any)[property] = (data as any)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.creator_id = _data["creator_id"] !== undefined ? _data["creator_id"] : null as any;
            this.time = _data["time"] ? new Date(_data["time"].toString()) : null as any;
            this.value = _data["value"] !== undefined ? _data["value"] : null as any;
        }
    }

    static fromJS(data: any): VoteDto {
        data = typeof data === 'object' ? data : {};
        let result = new VoteDto();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["creator_id"] = this.creator_id !== undefined ? this.creator_id : null as any;
        data["time"] = this.time ? this.time.toISOString() : null as any;
        data["value"] = this.value !== undefined ? this.value : null as any;
        return data;
    }
}